[ST40] Add support for READ-ONLY Boot-from-RAM capability
This commit will add initial support to allow U-Boot to be used
as a Secondary Boot Loader (SBL), in a "READ-ONLY boot-from-RAM"
environment. When booting from RAM, U-Boot is essentially only a SBL,
and some other "entity" has performed the duty of a primary boot loader
(PBL, or another booting mechanism). In this case, U-Boot does not need
to perform the pokes in the "poke-table", or even relocate itself.

The READ-ONLY aspect is that when U-Boot starts, the "u-boot.bin"
image has already been loaded into RAM (by some other entity), but that
RAM has been set to be READ-ONLY by the underlying hardware. In this
case, the ".text" (and other read-only) sections can be used directly
"in-place", without any more copying/relocation. Whereas, the ".data"
(and other read-write) sections need to be re-relocated (from RO-RAM to
RW-RAM) very early in U-Boot's initialization. It is this specific aspect
(boot-from-RAM in a READ-ONLY environment) that this commit is addressing.

Essentially, the READ-WRITE data sections need to be linked in a special
way so that they have virtual memory addresses (VMA) that differ from
their load memory addresses (LMA). Then U-Boot, will need to "re-relocate"
these sections (from RO-RAM to RW-RAM).

With this new capability, the main "READ-WRITE" sections are represented
by the following memory map of VMA:

                  ---------------
0x00000000        ...
                  ---------------
                  STACK (...)
                  STACK (growth downwards)
                  STACK (...)
                  STACK (start)
                  ---------------
__USE_VMA_DATA -> .data (start)
                  .data (...)
                  .data (end)
                  ---------------
                  .u_boot_cmd (start)
                  .u_boot_cmd (...)
                  .u_boot_cmd (end)
                  ---------------
                  .bss (start)
                  .bss (...)
                  .bss (end)
                  ---------------
0xffffffff        ...
                  ---------------

By default, this capability is NOT enabled, and so there is no change to
the usage of U-Boot. Users who wish to use this capability will have to
decide on their own specific memory map for U-Boot, and configure U-Boot
accordingly. Some of the steps necessary, are:

	1) Edit the file "cpu/st40/u-boot.lds" and set the linker symbol
	   "__USE_VMA_DATA" to be the VMA where the read-write ".data"
	   sections are to be copied. Setting __USE_VMA_DATA to zero,
	   disables this capability.

	2) Define the following macro in the board-specific configuration
	   file:

		#define CONFIG_SYS_USE_VMA_DATA

	3) Create an empty ".romgen" file (i.e. no pokes).  Alternatively,
	   you can define the following two macros to reference a 12-byte
	   region of ZEROS, which will yield an empty poke-table.

		#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_BEGIN
		#define CONFIG_SYS_ABSOLUTE_POKE_TABLE_END

	4) Verify the location of the STACK, that it is still safe and
	   appropriate.  By default, with this capability enabled,
	   the stack will start at the value of __USE_VMA_DATA,
	   and grow downwards. (The first PUSH will be a write to
	   __USE_VMA_DATA-0x4).  Hence, stack underflow would corrupt the
	   ".data" section! It may be changed by modifying the value at
	   the label "stack_addr" in the file "cpu/st40/start.S".

	5) Verify the location of the HEAP, that it is still safe and
	   appropriate. By default the HEAP is defined as having a
	   length of CONFIG_SYS_MALLOC_LEN bytes, and grows upwards,
	   ending at TEXT_BASE.

	6) Verify the location of the GLOBAL-DATA, that it is still safe
	   and appropriate. By default the GLOBAL-DATA is defined as
	   having a length of CONFIG_SYS_GBL_DATA_SIZE bytes, and ends
	   just below the HEAP.

	7) Verify the bounds used by the "mtest" command, that it
	   is still safe and appropriate.  The bounds are defined
	   by CONFIG_SYS_MEMTEST_START and CONFIG_SYS_MEMTEST_END.
	   In general, it is essential that CONFIG_SYS_MEMTEST_END is
	   below the .bss, .data, STACK, HEAP and GLOBAL-DATA regions.

	8) (Optionally) modify the address U-Boot starts at, by modifying
	   the value of TEXT_BASE, typically in the top-level "Makefile".

See the GIT commit logs of the last few commits for further details,
as well as comments in the following files:
		cpu/st40/init_ram.S
		cpu/st40/start.S
		cpu/st40/u-boot.lds

This have been given very little testing, but crucially, with these
features disabled (the default), then there is no logical change in
U-Boot's behaviour.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>