[ST40,FIX] Allow boot-from-FLASH to work when __USE_VMA_DATA!=0
This is a simple, but important, bug-fix to the previous commit.

Whilst testing boot-from-FLASH, with __USE_VMA_DATA!=0 (i.e. requiring
the ".data" section to be re-relocated - RAM to RAM), it was notice that
the relocation of u-boot.bin always assumed the destination region was
defined by ".text" to ".bss".  This is correct when __USE_VMA_DATA==0
(the long-held (implicit) default), but incorrect otherwise!

We now use a new location "target_end" just to store the end of the
".data" region, and now exclusively use the location "bss_start" to store
the start of the ".bss". Previously (to save space) "bss_start" was (quite
legitimately) used for both purposes, but this is no longer valid.  Note:
when __USE_VMA_DATA==0 these two locations will hold identical addresses.

In addition, we also provide some putative (untested) code to deal with
the case when the BART may be active, and we may be running out of the
BART, and hence we do not need to relocate u-boot.bin, even though the PC
"appears" to be in the wrong location (!=TEXT_BASE), that would normally
be indicative of (boot-from-FLASH, and hence) a copy of the whole of
u-boot.bin from FLASH to RAM being necessary. So, if we are running out
of the BART, then we do not perform an unnecessary copy of u-boot.bin.

Testing of this (modulo the BART code) was successful with
boot-from-FLASH, both with and without __USE_VMA_DATA == 0. That is
with and without ".data" being re-relocated when booting from NAND.
This was successfully tested with no issues noted, on a single HDK-7108
board (STi7108-HDK22ST-E835).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>