[ST40] Enable a few more generally useful commands ...
The following #define macros are now defined in include/asm-st40/config.h,
but are compiled out (with "#if 0 ... #endif") by default:

	CONFIG_CMD_STTESTS	/* provides "st_mtest" */
	CONFIG_CMD_MD5SUM	/* provides "md5sum" */
	CONFIG_CMD_SHA1		/* provides "sha1sum" */
	CONFIG_CMD_STRINGS	/* provides "strings" */
	CONFIG_CMD_UNZIP	/* provides "unzip" */
	CONFIG_MD5		/* needed by CONFIG_CMD_MD5SUM */
	CONFIG_SHA1		/* needed by CONFIG_CMD_SHA1 */

This can easily enable a small set of additional commands that may prove
useful, but perhaps should not always be present, in order to keep normal
code size to a minimum.  To enable them, the "#if 0" needs to be changed
to "#if 1" (or similar).

In addition, a trivial bug fix (for a minor typo) in "common/cmd_mem.c"
is also included here, otherwise the "sha1sum" command will not actually
be usable.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>