[ST40] Add initial support for the B2075 Board (STxH205)
This commit adds initial support for the B2075 board from
STMicroelectronics, which is also known as the STiH205-HDK board, which
has a STiH205 SoC device.

This board is very similar to previously supported boards (the B2057,
B2064, and the B2067).  It is so similar that it was decided to reuse
the existing board/st/b2057 directory and the files therein for it, and
not create a separate directory (board/st/b2075), as one would normally
do to support a new board.

Only basic testing of this support has been done, using:

	B2075A board (B2075A-00-R036)

Testing has included support for: serial console (on-board CN32,
and off-board JD7); Networking (with in-package IC+ IP101G PHY); USB
(OHCI and EHCI); and multi-bit ECC NAND (MT29F4G08ABADAH4) using the
BCH controller in 18-bit ECC mode; and the "reset" command successfully
re-boots the board.  Boot-from-NAND has been tested, and the board was
able to boot from NAND with no problems (with BCH18).

NAND was only tested with 18-bit ECC on the BCH controller, as the fitted
NAND device is not capable of supporting the ECC needs for 30-bit ECC,
and using a weaker ECC is not recommended!

USB was tested successfully using both ports (USB #0 on JD1, and USB #1
on JD3) with both the OHCI (USB 1.x), and the EHCI (USB 2.0) controllers.

It should be noted, that this commit does *not* include support for SPI
Serial FLASH. The specific SPI device fitted (in UD15/UD16) on the B2075A
is not a device that has previously been used by STMicroelectronics, hence
there is no existing driver for it presently in the STMicroelectronics
ST40 tree. It should also be noted that the SPI device fitted is
different from the ones suggested in the schematics for the B2075A.
The addition of a suitable driver, and hence support for the specific
SPI Serial FLASH device fitted should be added soon...

Note: Please ensure you read the comments near the beginning of the
"include/configs/b2075.h" file for details about setting up the mode pins
correctly. The ones in this file have been empirically verified for the
specific NAND device on the board (MT29F4G08ABADAH4), which is a 512 MiB
NAND, with a page size of 2048+64 bytes and a sector size of 128 KiB,
requiring 5 address cycles. If you are using a different NAND device,
then you may need to modify the mode-pins to boot from it - check the
SoC's data-sheet for details.

Note: It is recommended that the romgen utility is re-run and the file
"board/st/b2057/b2075-stxh205.romgen" is replaced, when building
an image to boot from FLASH. Please read the comments in the file
"board/st/b2057/init-b2075.S" for details on how to run romgen.

Finally, this commit also includes putative confirmation that the recent
re-base of the ST40 tree from release 'v2009.03' to release 'v2010.03'
appears to be successful, and should be logically completed - albeit
only tested on a single ("new") board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>