[ST40,BCH,FIX] Dynamically select 4 or 5 address cycles for BCH NAND
Prior to this fix, U-Boot on ST40 unconditionally assumed that the NAND
device, when used with the BCH controller, always used 5 address cycles.
However, there are several suitable NAND devices which require only 4
address cycles, and these were not supported correctly.

This commit, fixes this bug, and will now dynamically select the correct
number of address cycles to use (4 or 5), depending on the specific NAND
device that was probed.

Unfortunately, I do not have access to any boards with a NAND device
requiring only 4 address cycles, so this was not tested. A quick test was
performed on a single ST40 board with a NAND requiring 5 address cycles,
which had previously been known to work for some time.

See also:
	https://bugzilla.stlinux.com/show_bug.cgi?id=22697

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>