[ST40] Add "helper" functions to read SPI FLAG STATUS register
This commit adds a couple of "helper" functions, to read the "FLAG STATUS"
register on SPI Serial Flash devices, for ST40 SoCs.

Specifically, for the N25Q256 SPI device, there is a "FLAG STATUS"
register, which contains a bit which indicates if the SPI device is in
the 3-byte (24-bit), or the 4-byte (32-bit) addressing mode. Knowing the
actual value of this bit is useful when debugging code which is trying to
set/clear it! Hence, these functions (currently) are intended primarily
for "debugging" purposes.

The 2 new functions are:

	spi_read_flag_status(), and fsm_read_flag_status().

They both return the current value of the flag status register.

By default these functions are compiled-out. So, to compile them in,
both the following macros should be defined:

	#define CONFIG_SPI_FLASH_ST
	#define CONFIG_STM_SPI_READ_FLAG_FUNCTION

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>