[ST40] Add 32-bit addressing with SPI on H/W FSM (both READ+WRITE)
This commit completes (initial) support for READ+WRITE (and ERASE)
operations for SPI Serial Flashes with 32-bit (4-byte) addressing,
required for SPI devices > 16MiB, (such as the N25Q256 device), using
ST40 SoCs.

Specifically this commit adds support using the H/W FSM SPI Serial Flash
controller on recent ST40 SoCs - my recent commits were not for the H/W
FSM driver, but for S/W "bit-banging", and H/W SSC drivers.

Tested on a single B2064B board (STiH239-HDKB200 E075) with an on-board
N25Q256 SPI Serial Flash, using the H/W "FSM" driver.

With this commit, it should now be possible to access all of the 32MiB
on the N25Q256 SPI serial flash device, using any of the bit-banging,
SSC, or FSM drivers in U-Boot.

Note: The current implementation for dealing with 32-bit addresses
using the FSM relies on direct H/W support for 32-bit addressing, which
is only included in recent SoCs from STMicroelectronics. Specifically,
such H/W support is included in the STxH205, and STXH415 SoCs - but not
in older SoCs such as the STx7105, nor the STx7108.

The code will "BUG()" if an attempt is make to use 32-bit addressing
on SoCs which do not support it. However, that should not happen, as
the driver will "revert" instead to using only 24-bit addresses, and
effectively cap the device at the first 16MiB. In which case ,diagnostics
similar to the following may then be seen:

SPI:  info: found ST N25Q256 (32MiB) device (page=256,erase=65536)
warning: SPI FSM controller does not support 32-bit addresses!
warning: SPI device (ST N25Q256) effectively capped at first 16MiB only!

Should there be a need to support 32-bit addressing on SPI devices
with the (older) H/W FSM controller, which does not natively support
32-bit addresses, then additional work to the U-Boot driver will be
required. Probably, a "pair" of 16-bit address will be sent with 2
"ADDR" commands. As a short-term workaround to use 32-bit addresses
on SoCs which do not support this capability on the FSM, then the S/W
"bit-banging" driver, or the H/W SSC driver may be used, as appropriate.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>