[ST40] Add initial WRITE support for Spansion SPI (S25FLxxxS)
This commit adds initial (write+erase) support for the Spansion S25FLxxxS
family of SPI Serial Flash Devices, used on the STMicroelectronics' SoC
devices. In particular, this is the second of a series of incremental
patches to provide support for both the S25FL128S and the S25FL256S
SPI devices.

To add support for these Spansion SPI Serial Flash devices, the following
macro should be defined:

	#define CONFIG_SPI_FLASH_SPANSION

This code will use the (new) 4-byte WRITE ("4PP", 0x12) command, and the
(new) 4-byte ERASE command ("4SE", 0xDC) for devices greater than 16MiB,
as well as continuing to use the 4-byte READ ("4READ", 0x13) command used
in the previous commit.  As a result, the driver will *not* enter 32-bit
address mode (ever!), and hence remain in the default 24-bit address
mode, which allows asynchronous resets still to boot from SPI correctly,
without having to worry about "synchronization" issues between the SPI
controller, and the SPI device. Although, there may still be an issue
if WIP is still active when the CPU tries to boot-from-SPI after a reset!

In order to facilitate this new support, the code was re-factored
slightly, to allow the 3 SPI commands for READ, WRITE and ERASE, to be
overridden easily, and consistently, depending on the SPI device that
is probed.

Note: Spansion SPI serial flash devices with a uniform sector size of
      256KiB are not supported, and currently, an erase sector size of
      64KiB will always be assumed. Additional device probing will be
      required to differentiate between the real sector sizes present.

Only basic testing of this new support has been done, using a single
B2075A board (B2075A-00-R036), with a (bottom) S25FL256S SPI device.
As yet, write support has not been tested, as all the sectors are
automatically "locked" on power-on, hence un-writeable!  So it was
not possible to program any data patters for testing purposes, hence
the "read" testing just yielded 0xFF bytes only! However, testing was
performed using both the FSM driver and "bit-banging" driver, (but not
with the SSC driver).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>