[ST40] Add DYB unlock support for Spansion SPI (S25FLxxxS)
This commit adds (DYB) Sector Locking/Unlocking support for the
Spansion S25FLxxxS family of SPI Serial Flash Devices, used on the
STMicroelectronics' SoC devices. In particular, this is the third of a
series of incremental patches to provide support for both the S25FL128S
and the S25FL256S SPI devices.

This commit adds new functions to READ and WRITE the "DYB Access
Register" for each individual sector, on the aforementioned SPI Serial
Flash devices. This new capability has been implemented with both the
FSM driver, as well as the "bit-banging", and the SSC drivers.

In addition, a new function "spi_dby_unlock_all()" will unlock all the
sectors protected by the DYB protection mechanism, on Spansion devices.
Recall, for the Spansion S25FLxxxS family of parts, there are three main
sector topologies: "UNIFORM", "BOTTOM", or "TOP".

	UNIFORM: all sectors are of the same size, being 256KiB.

	BOTTOM:  the first 32 sectors are all 4KiB, and the
	         remainder are all 64KiB.

	TOP:     the last 32 sectors are all 4KiB, and the remainder
	         are all 64KiB.

In order to lock/unlock all the sectors, then it is essential that each
sector is individually addressed. However, we play it safe, and assume
the device is a "union" of "TOP" and "BOTTOM", by unlocking every block
assuming a 64KiB stride, and also unlocking the top 32 and bottom 32 with
a 4KiB stride.  This results in doing more that than strictly necessary,
but it means we do not need to know the exact topology, but we can be
confident of unlocking *all* of the sectors.

Only minimal testing of this new support has been done, using a single
B2075A board (B2075A-00-R036), with a (bottom) S25FL256S SPI device.
All testing was performed using both the H/W FSM driver and the S/W
"bit-banging" driver, (but not with the SSC driver). Various data
patterns were written to the SPI, and read back correctly.  In addition,
boot-from-SPI on the B2075A was also successful (both FSM + bit-banging).

On first execution after a "reset", we now get:

	info: found Spansion S25FL256S (32MiB) device (page=256,erase=65536)
	info: Unlocked 542 sectors with DYB on Spansion S25FL256S

On subsequent executions (no additional "reset"), we now get:

	info: found Spansion S25FL256S (32MiB) device (page=256,erase=65536)

That is, the DYB unlocking survives across multiple GDB connects, hence
only the first execution after a reset actually unlocks the sectors.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>