[ST40] Prepare for removal of CONFIG_ST40_OLD_POKE_LOOP_PARADIGM
This commit is preparatory, to permit us to remove the ability to use
the OLD "poke interpreter". This commit does not remove the old poke
interpreter from the source tree, it simply stops using it, and converting
all such use cases to use the NEW poke interpreter instead.

No testing was done, expect to ensure that the boards which had affected
files, were still able to be compiled, with no additional diagnostics.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>