[ST40] Remove CONFIG_ST40_OLD_POKE_LOOP_PARADIGM
Following on from the previous commit, this commit removes the OLD "poke
interpeter", and all referecnes to it, and all backward compatabilies.
Specifically, there is now only one poke interpreter, and "romgen"
should be used to generate the final "poke-table" to be used (i.e. the
".romgen" files).

No testing was done, expect to ensure that all the ST40 boards, were
still able to be compiled, with no additional diagnostics raised.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>