[ST40] Adds new (top-level) make target of "romgen"
This commit adds a new top-level make target of "romgen" (i.e. "make
romgen"), which is used to run the poke-table generator utility "romgen",
from STMicroelectronics.

The new target should be exercised *after* the configuration stage,
but before the building stage. That is:

	host% make distclean
	host% make <board>_config
	host% make romgen
	host% make

The new target will run the "romgen" utility to create a replacement
".romgen" file in situ. In general, the ".romgen" files included in the
U-Boot release are never updated, and hence it is strongly recommended
that romgen is always re-run by the user to get the latest set of
"pokes". That policy has not changed, what has changed is there is now
a new and easy way for users to run romgen automatically, integrated as
part of U-Boot "make" system.

The general philosophy is to use romgen in a mode where it does *not*
need to connect to a real target board. For most of the newer ST40 system,
this should not be a problem. For some of the older ST40 systems, then
"romgen" may fail to generate a "robust" file. However, it is possible
to pass additional parameters, or over-ride the parameters/options passed
to romgen, which usually allows romgen to succeed.

To change the behaviour of "romgen", then the following make environment
variables can be used to control how the romgen utility is invoked:

ROMGEN_STMC2_NAME		DNS name of STMC2 to use, or the special
				string "no-such-ip", if no *physical*
				board is to be used.  It is generally
				recommended to use "no-such-ip".

ROMGEN_TARGETPACK_NAME		The name of the TargetPack to use.

ROMGEN_CORE_NAME		The name of the specific core to use.

ROMGEN_DEFAULT_PARAMETERS	A comma-separated list of the DEFAULT
				parameters to be passed to romgen.

ROMGEN_EXTRA_PARAMETERS		A comma-separated list of the ADDITIONAL
				parameters to be passed to romgen.

ROMGEN_EXTRA_OPTIONS		A space-separated list of the ADDITIONAL
				command-line options to be passed
				to romgen.

Note: the first 3 environment variables above are mandatory, and must be
defined - "make <board>_config" automatically does this for most boards.

To just "add" a few parameters, the simplest means is just to set the
ROMGEN_EXTRA_PARAMETERS parameter, e.g.:

	host% make romgen ROMGEN_EXTRA_PARAMETERS="se=0,lmi_freq=266"

To get romgen to connect to a real target board (via a STMC2), then
simply define ROMGEN_STMC2_NAME to be the DNS name of the STMC2. e.g.:

	host% make romgen ROMGEN_STMC2_NAME=mystmc2

The resultant romgen command is (roughly) constructed as follows:

host% romgen
	--output="board/st/${BASE_BOARD}/${SPECIFIC_BOARD}-${SOC}.romgen"
	--show-comment
	--warn-peek OR --target-connect		(only one of them)
	${ROMGEN_EXTRA_OPTIONS}
	--
	${ROMGEN_STMC2_NAME}:${ROMGEN_TARGETPACK_NAME}:${ROMGEN_CORE_NAME},${ROMGEN_DEFAULT_PARAMETERS},${ROMGEN_EXTRA_PARAMETERS}

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>