[ST40] Update romgen comments in board/st/*/init-*.S
This code logically changes nothing, inasmuch as it just updates some
comments (in board/st/*/init-*.S) to clarify which parameters are required
to be passed to romgen, for the set of boards where "make romgen" should
work, without the need to connect to a real target board.

No testing was done, expect to ensure it still compiles.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>