[ST40] Re-base pokeloop interpreter to version 4.1
This code essentially updates the STMicroelectronics' "pokeloop"
interpreter to the latest version, which is currently version 4.1

The update is being done primarily to support a new feature added
in the latest version of "romgen" (2012.2.1). Specifically the new
"--batch-pokes" command-line option, which previously U-Boot did not
support. This now allows for the creation of more compact (smaller)
poke-tables.

Essentially, this is just importing the two files "poketable.h" and
"sh4/pokeloop.S", from STMicroelectronics' "poketable.git" repository,
which were tagged with "pokeloop_4.1", into the U-Boot repository.

A few minor "infrastructure" changes were also required due to this
re-baselining activity, to ease the integration, and to make the resultant
u-boot.bin as similar as possible to previous ones.

In addition, some deprecated opcodes (#define macros) were deleted from
the new "poketable.h" file, to *remove* any backwards compatibility
"with very old poke tables". This forces users to use contemporary files
(or to re-run the romgen utility) - which really, is for the best! The
deleted opcodes are:

	POKE_CHAR, POKE_SHORT, POKE_LONG, OR_LONG, UPDATE_LONG,
	POKE_UPDATE_LONG, POKE_UPDATE32 and WHILE_NE.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>