[ST40] Big bunch of updates for STxH415 and B2000 board
This commit prepares the integration of Orly in U-Boot.
This include:
 - romgen update
 - ethernet tweak and fixes
 - usb tweak and fixes
 - ... and testing!

The original patch was provided by Jerome Brunet, and was reviewed, and
edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Jerome for the original patch (originally, 1 of 20).

Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>