[STM] Adds "generic" PHY capability for the STMAC driver.
This commit adds a "generic" PHY capability to STMicroelectronics'
stmac ethernet driver, which provides a minimal level of support,
and configuration.

To use it, the following macro should be defined:
	#define CONFIG_STMAC_GENERIC

The original patch was provided by Jerome Brunet, and was reviewed, and
edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Jerome for the original patch (originally, 1 of 20).

Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>