[ST40] Eschew bash 4.x specific features from "romgen-wrapper"
In bash, the token "|&" was added for version 4.x, as a convenient
abbreviation for the oft used idiom "2>&1 |", to redirect both stdout
and stderr into the pipe-line.

However, some users are still using older linux systems deploying
earlier versions of bash (e.g. 3.x), and this results in the script
"romgen-wrapper" on such (older) systems failing!

This commit is exclusively to use the long-established idiom, which
will work equally well on both older and never versions of the bash
shell. Essentially, we are just adding backwards-compatibility for bash,
and avoid using bash 4.x specific features.

Tested by running "make romgen", albeit on a system with bash 4.2.28.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>