Increase command buffer to handle longer command-lines
Increase MAX_CMDBUF_SIZE (the command-line buffer, from 256) so we can
define even longer U-Boot environment variables (in particular "bootargs")
which can nowadays grow to be quite big!

The original patch was provided by Jerome Brunet, and was reviewed, and
edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Jerome for the original patch (originally, 5 of 20).

Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>