[ST40] add roundup() definition in include/asm-st40/config.h
In the include/configs/<board>.h files, we may want to use the roundup()
macro, to ensure that various configuration #defines are appropriately
"rounded" up. This is especially true for dealing with very-large page
NAND devices, and the definitions of both the CONFIG_ENV_SIZE, and the
CONFIG_SYS_NAND_ENV_OFFSET macros, which typically will be multiples of
NAND erase-size blocks.

We include this macro definition in include/asm-st40/config.h, to avoid
its definition being replicated in multiple ST40 configuration files.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>