[ST40] Add Marvell PHY support to ST-GMAC for customer boards
This patch adds support for the Marvell (88EC060) PHY, to the
STMicroelectronics' GMAC driver, in U-Boot, primarily for some (anonymous)
"customer" boards.

The original patch was provided by Jerome Brunet, and was reviewed, and
edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Jerome for the original patch (originally, 7 of 20).

Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>