[ST40] Add support for the MXIC MX25L3206E SPI FLASH
This commit adds support for the MXIC MX25L3206E SPI serial FLASH device,
to the ST40 SPI serial flash driver in U-Boot.

The original patch was provided by Jerome Brunet, and was reviewed, and
edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Jerome for the original patch (originally, 10 of 20).

Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>