[ST40] First delivery of support for STxH416 (Orly2)
This commit, is the initial delivery to add support to U-Boot for the
STxH416 (aka Orly2) SoC from STMicroelectronics, on both the B2000 and
the B2020 reference boards from STMicroelectronics.

There are now 18 make targets for the B2000 and the B2020 boards:

	b2000se_config
	b2000se_spi_config
	b2000se_nand_config

	b2000stxh415se_config
	b2000stxh415se_spi_config
	b2000stxh415se_nand_config

	b2000stxh416se_config
	b2000stxh416se_spi_config
	b2000stxh416se_nand_config

	b2020se_config
	b2020se_spi_config
	b2020se_nand_config

	b2020stxh415se_config
	b2020stxh415se_spi_config
	b2020stxh415se_nand_config

	b2020stxh416se_config
	b2020stxh416se_spi_config
	b2020stxh416se_nand_config

Please note: if "_spi", or "_nand" is not explicitly given, then U-Boot
will default to boot-from-SPI.  Also, if "stxh415", or "stxh416" is not
explicitly given, then U-Boot will default to targeting the STxH415 SoC.

The original patch was provided by Marc Chappellier, and was reviewed,
and edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Marc for the original patch (originally, 13 of 20).

Signed-off-by: Marc Chappellier <marc.chappellier@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>