[ST40] Fix ethernet initialization on the STxH416
This commit uses corrected SYSCONF register addresses/bitfields,
specifically for the STxH416 SoC from STMicroelectronics. It was tested
(by the original author) on a B2000 board.

The original patch was provided by Marc Chappellier, and was reviewed,
and edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Marc for the original patch (originally, 15 of 20).

Signed-off-by: Marc Chappellier <marc.chappellier@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>