[ST40] Prepare for booting ARM images on STxH416
This commit uses corrected SYSCONF registers, specifically for the STxH416
(Orly2) SoC from STMicroelectronics. This is in order to facilitate the
ST40 core to boot ARM images on this SoC.

The original patch was provided by Marc Chappellier, and was reviewed,
and edited by Sean McGoogan, prior to being committed. At time of commit,
Sean McGoogan had not tested it on any board, but only compiled it.
Thanks to Marc for the original patch (originally, 16 of 20).

Signed-off-by: Marc Chappellier <marc.chappellier@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>