[STM] New romgen files for CA9 at 1000MHz and 1200MHz
Create a set of 4 romgen files for the CA9 running both at 1000MHz,
and 1200MHz; and using two versions of the TargetPacks (v9, and v12).

The original patch was provided by Nicolas Chesné, and was reviewed by
Sean McGoogan, prior to being committed. At time of commit, Sean McGoogan
had not tested it on any board, but only compiled it.  Thanks to Nicolas
for the original patch (originally, 17 of 20).

Signed-off-by: Nicolas Chesné <nicolas.chesne@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>