[ST40] Use AHB2STBUS_OHCI_BASE as base for OHCI (USB) registers.
For STMicroelectronics' systems, we will now define the OHCI base address
macro (CONFIG_SYS_USB_OHCI_REGS_BASE), in terms of AHB2STBUS_OHCI_BASE,
which is defined in the ST40-specific file "stbus.h".

This allows the offset to be defined based on SoC-specific files, and not
a board-specific file. Historically the OHCI registers were at a fixed
offset of 0xffc00 from CONFIG_SYS_USB_BASE. However, on the STxH416
(Orly2), this offset was changed. As both the B2000 and B2020 boards
can use either STxH415 (Orly1) or STxH416 (Orly2) SoCs, then it is now
inappropriate to select the actual OHCI offset in the board-specific
configuration file, and it is better to have this defined in some "common"
file, to reduce replication. This is the prime rationale for this commit.

Note this has not been tested, apart from compilation. However, logically,
nothing should have changed.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>