[ST40] Use revised USB offsets for the STxH416 (Orly2)
Prior to the STxH416 (Orly2), all ST40 SoCs from STMicroelectronics
reserved (in the memory-map) 1MiB for *each* USB IP block instantiated.
Hence, the register addresses were all aligned with, and constrained
within, this 1MiB region.

However, the STxH416 (Orly2) has started to limit this memory-mapped
region for USB to only 256KiB, and hence all the old registers that are
now outwith the (newer) smaller region, have been moved downwards.

This commit moves these registers for the STxH416 (Orly2) to fit within
the new smaller 256KiB region, which is required for correct operation
of these on-die USB blocks!

Note: any new SoC with the same offsets as the STxH416 will need to
expand the '#if defined()' line to use the correct offsets!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>