[ST40] Rationalize the "re-timing" structures for STxH416 (Orly2)
This commit rationalizes the "re-timing" structures, specifically aimed to
support the newer STxH416 (Orly2) SoC, whilst still maintaining support
for previous SoCs.  The STxH416 uses 4-bits for delays (previously only
2 bits), and 2-bits for clocks (previously only 1 bit).

Specifically, in the structure "stm_pio_control_retime_config", the old
bitfield "clk1notclk0" has now been renamed to "clk", and has grown by
1 bit, now to store a 2-bit value. Also, the old bitfield "delay_input"
has been renamed to "delay", and has grown by 2 bits, now to store a
4-bit value.  This growth is a direct result of the new PIO re-timing
register layout on the STxH416 (Orly2) SoC. However, the same (modified)
structure is still compatible with the re-timing register layouts for
earlier SoCs, with the additional bits never being used.

Note: With this code, each of the individual bitfields
still has one "extra" bit (a signed-bit) in the aforementioned
stm_pio_control_retime_config structure, which is used to allow us to set
the (signed) bitfield to a negative number (-1).  The use of negative
numbers allows us to convey an additional semantic of "do nothing" (or
"don't care"), such that we do not alter the current state of the H/W
bitfield. Use of this semantic is now deprecated, and will be removed in
the future (and the bitfields made unsigned). Better determinism can be
obtained by explicitly setting *all* bitfields unconditionally, as this
means that the "previous state" of a H/W bitfield is irrelevant to the
new (desired) state.  Currently, we still support the old semantic of -1
(to do nothing).

This has not been tested, apart from compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>