[ST40] removal of "don't care" (-1) semantics of re-timing structure.
This commit removes any "don't care" (-1) semantics from all the
RET_xxx() retiming macros, which are used to initialize an instance of
the "stm_pio_control_retime_config" structure.

Historically, if any of the bitfields in the stm_pio_control_retime_config
structure were negative (usually -1), then that bitfield would have
the semantics of "don't care", that is the H/W bitfields would not be
modified. This commit replaces any default, or explicit initialization
of a negative value, with a positive one (well zero). The result is that
*all* bitfields should now be explicitly initialized, and not rely on
the current (or previous) state of the H/W bitfields to be correct.

Note: the SoC-specific code will still honour the old semantics of
negative numbers meaning "don't care". However, this usage is now
deprecated, and should be removed in the future. This commit simply
stops the above bitfields being initialized with a negative value.

This has not been tested, apart from compilation.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>